[Respects] Cog model restructure
Restructure cog `Respects` according to SFUAnime/Ren#608. No functional
changes.
